Cleanup of various parts related to account->player handling.
