+++ client/osci/dialog/NewAccountDlg.py
@@ -20,117 +20,123 @@
 
 import pygameui as ui
 from osci import client, gdata, res
-from ige.ospace.Const import *
 from MainGameDlg import MainGameDlg
-from ConfirmDlg import ConfirmDlg
-from BookingDlg import BookingDlg
-import ige
+from ige import SecurityException
 
 class NewAccountDlg:
-    """ Called for a new player."""
 
     def __init__(self, app):
         self.app = app
         self.createUI()
-        self.confirmDlg = ConfirmDlg(app)
-        self.confirmDlg.setTitle(_("No free starting position"))
 
-    def display(self, caller = None):
+    def display(self, caller = None, message = None):
         self.caller = caller
-        if self.show():
-            self.win.show()
+        self.win.vNick.text = ""
+        self.win.vLogin.text = ""
+        self.win.vPassword.text = ""
+        self.win.vConfirm.text = ""
+        self.win.vEmail.text = ""
+        self.win.vMessage.text = message
+        self.win.show()
 
     def hide(self):
         self.win.hide()
 
-    def show(self):
-        positions = client.cmdProxy.getStartingPositions()
-        items = []
-        for objID, name, posType in positions:
-            item = ui.Item(name, tObjID = objID, tPosType = posType)
-            if posType == STARTPOS_NEWPLAYER:
-                item.tPos = _('Independent player')
-            elif posType == STARTPOS_AIPLAYER:
-                item.tPos = _("Rebel faction")
-            elif posType == STARTPOS_PIRATE:
-                item.tPos = _("Pirate faction [VIP password needed]")
-            else:
-                item.tPos = _('Unknown. You cannot use this.')
-            items.append(item)
-        self.win.vPos.setItems(items)
-        return True
-
-    def onSelect(self, widget, action, data):
-        if not self.win.vPos.selection:
-            self.win.setStatus(_('Select position.'))
+    def onCreate(self, widget, action, data):
+        nick = self.win.vNick.text
+        login = self.win.vLogin.text
+        password = self.win.vPassword.text
+        confirm = self.win.vConfirm.text
+        email = self.win.vEmail.text
+        if password != confirm:
+            self.win.vMessage.text = _("Passwords do not match.")
+            self.win.vPassword.text = ""
+            self.win.vConfirm.text = ""
             return
-        item = self.win.vPos.selection[0]
-        if item.tPosType == STARTPOS_NEWPLAYER:
-            self.win.setStatus(_('Executing CREATE NEW PLAYER command...'))
-            playerID = client.cmdProxy.createNewPlayer(item.tObjID)
-            self.win.setStatus(_('Command has been executed.'))
-            self.hide()
-            if not gdata.mainGameDlg:
-                gdata.mainGameDlg = MainGameDlg(self.app)
-                gdata.mainGameDlg.display()
-            client.updateDatabase(clearDB = 1)
-        elif item.tPosType == STARTPOS_AIPLAYER:
-            self.win.setStatus(_('Executing TAKE OVER REBEL FACTION command...'))
-            playerID = client.cmdProxy.takeOverAIPlayer(item.tObjID)
-            self.win.setStatus(_('Command has been executed.'))
-            self.hide()
-            if not gdata.mainGameDlg:
-                gdata.mainGameDlg = MainGameDlg(self.app)
-                gdata.mainGameDlg.display()
-            client.updateDatabase(clearDB = 1)
-        elif item.tPosType == STARTPOS_PIRATE:
-            password = self.win.vPassword.text
-            if not password:
-                self.win.setStatus(_("Supply VIP password, please."))
-                return
-            self.win.setStatus(_('Executing TAKE OVER PIRATE FACTION command...'))
-            try:
-                playerID = client.cmdProxy.takeOverPirate(item.tObjID, password)
-            except ige.SecurityException:
-                self.win.setStatus(_("Supply valid VIP password."))
-                return
-            self.win.setStatus(_('Command has been executed.'))
-            self.hide()
-            if not gdata.mainGameDlg:
-                gdata.mainGameDlg = MainGameDlg(self.app)
-                gdata.mainGameDlg.display()
-            client.updateDatabase(clearDB = 1)
-
-    def onBooking(self, widget, action, data):
         self.win.hide()
-        self.win.setStatus(_('Command has been executed.'))
-        BookingDlg(gdata.app).display(self)
+        try:
+            result = client.createAccount(login, password, nick, email)
+        except SecurityException, e:
+            # failed
+            self.win.vMessage.text = _(e.args[0])
+            self.win.show()
+            return
+        # account created, record login
+        gdata.config.game.lastlogin = login
+        gdata.config.game.lastpassword = ""
+        gdata.config.game.accountcreated = 1
+        if self.caller:
+            self.caller.display(message = "The account has been created.")
 
     def onCancel(self, widget, action, data):
         self.win.hide()
         if self.caller:
             self.caller.display()
-        else:
-            self.app.exit()
 
     def createUI(self):
         w, h = gdata.scrnSize
         self.win = ui.Window(self.app,
             modal = 1,
             movable = 0,
-            title = _('Select starting position'),
-            rect = ui.Rect((w - 424) / 2, (h - 264) / 2, 424, 264),
+            title = _('Create account'),
+            rect = ui.Rect((w - 424) / 2, (h - 144) / 2, 424, 144),
             layoutManager = ui.SimpleGridLM(),
             tabChange = True
         )
-        ui.Listbox(self.win, layout = (0, 0, 21, 10), id = 'vPos',
-            columns = ((_('Galaxy'), 'text', 5, ui.ALIGN_W), (_('Position'), 'tPos', 0, ui.ALIGN_W)),
-            columnLabels = 1)
         self.win.subscribeAction('*', self)
-        ui.Label(self.win, layout = (0, 10, 5, 1), text = _("VIP Password:"))
-        ui.Entry(self.win, layout = (5, 10, 5, 1), id = 'vPassword', align = ui.ALIGN_W, showChar = '*', orderNo = 1 )
-        ui.TitleButton(self.win, layout = (13, 10, 8, 1), text = _('Book position'), action = 'onBooking')
-        ui.Title(self.win, layout = (0, 11, 13, 1), id = 'vStatusBar', align = ui.ALIGN_W)
-        ui.TitleButton(self.win, layout = (13, 11, 4, 1), text = _('Exit'), action = 'onCancel')
-        ui.TitleButton(self.win, layout = (17, 11, 4, 1), text = _('Select'), action = 'onSelect')
-        self.win.statusBar = self.win.vStatusBar
+        ui.Label(self.win,
+            text = _('Nick'),
+            align = ui.ALIGN_E,
+            layout = (5, 0, 6, 1)
+        )
+        ui.Entry(self.win, id = 'vNick',
+            align = ui.ALIGN_W,
+            layout = (11, 0, 10, 1),
+            orderNo = 1
+        )
+        ui.Label(self.win,
+            text = _('Login'),
+            align = ui.ALIGN_E,
+            layout = (5, 1, 6, 1),
+        )
+        ui.Entry(self.win, id = 'vLogin',
+            align = ui.ALIGN_W,
+            layout = (11, 1, 10, 1),
+            orderNo = 2
+        )
+        ui.Label(self.win,
+            text = _('Password'),
+            align = ui.ALIGN_E,
+            layout = (5, 2, 6, 1),
+        )
+        ui.Entry(self.win, id = 'vPassword',
+            align = ui.ALIGN_W,
+            showChar = '*',
+            layout = (11, 2, 10, 1),
+            orderNo = 3
+        )
+        ui.Label(self.win,
+            align = ui.ALIGN_E,
+            text = _('Confirm'),
+            layout = (5, 3, 6, 1),
+        )
+        ui.Entry(self.win, id = 'vConfirm',
+            align = ui.ALIGN_W,
+            layout = (11, 3, 10, 1),
+            showChar = "*",
+            orderNo = 4
+        )
+        ui.Label(self.win,
+            align = ui.ALIGN_E,
+            text = _('E-mail'),
+            layout = (5, 4, 6, 1),
+        )
+        ui.Entry(self.win, id = 'vEmail',
+            align = ui.ALIGN_W,
+            layout = (11, 4, 10, 1),
+            orderNo = 5
+        )
+        ui.Title(self.win, layout = (0, 5, 13, 1), id = 'vMessage', align = ui.ALIGN_W)
+        ui.TitleButton(self.win, layout = (13, 5, 4, 1), text = _('Cancel'), action = 'onCancel')
+        ui.TitleButton(self.win, layout = (17, 5, 4, 1), text = _('Create'), action = 'onCreate')
+        ui.Label(self.win, layout = (0, 0, 5, 4), icons = ((res.loginLogoImg, ui.ALIGN_W),))
